add ci

+++ upload.js
@@ -47,7 +47,7 @@ module.exports = async ({github, context, core}) => {
     // Note: 从 push 事件中获取到相关文件变动信息，然后进行相应的上传和删除
     const addAndModifyList = [];
     const removedList = [];
-    const ignored = [];
+    const renamed = [];
     for (const file of files) {
         const filename = file.filename;
         switch (file.status) {
@@ -60,18 +60,17 @@ module.exports = async ({github, context, core}) => {
                 break;
             default:
                 // Note: 还有一种 renamed 的不处理
-                ignored.push({
+                renamed.push({
                     [file.status]: file
                 });
                 break;
         }
     }
-    console.log('未处理的文件有:', ignored);
-    // Note: 去重
     console.log('添加或修改的文件列表:', addAndModifyList);
     console.log('移除的文件列表:', removedList);
+    console.log('重命名的文件有:', renamed);
 
-    const addFiles = [];
+    const ignored = [];
     for (let fileName of addAndModifyList) {
         // Note: 只处理静态资源文件夹下的文件
         if (staticFileDir.some(dir => fileName.startsWith(dir))) {
@@ -85,10 +84,14 @@ module.exports = async ({github, context, core}) => {
                     console.log(JSON.stringify('进度:', processData.percent, processData.speed));
                 }
             });
+        } else {
+            ignored.push(fileName);
         }
     }
+    if (ignored.length) {
+        console.log('本次忽略上传的文件有:', ignored);
+    }
 
-    const removeFiles = [];
     for (let fileName of removedList) {
         if (staticFileDir.some(dir => fileName.startsWith(dir))) {
             console.log('即将删除文件:', fileName);
@@ -99,29 +102,37 @@ module.exports = async ({github, context, core}) => {
     }
 
     // Note: 执行上传
-    cos.uploadFiles({
-        files: addFiles,
-        SliceSize: 1024 * 1024 * 10,    /* 设置大于10MB采用分块上传 */
-        onProgress: function (info) {
-            var percent = parseInt(info.percent * 10000) / 100;
-            var speed = parseInt(info.speed / 1024 / 1024 * 100) / 100;
-            console.log('进度：' + percent + '%; 速度：' + speed + 'Mb/s;');
-        },
-        onFileFinish: function (err, data, options) {
-            console.log(options.Key + '上传' + (err ? '失败' : '完成'));
-        },
-    });
-
+    if (addAndModifyList.length) {
+        cos.uploadFiles({
+            files: addFiles,
+            SliceSize: 1024 * 1024 * 10,    /* 设置大于10MB采用分块上传 */
+            onProgress: function (info) {
+                var percent = parseInt(info.percent * 10000) / 100;
+                var speed = parseInt(info.speed / 1024 / 1024 * 100) / 100;
+                console.log('进度：' + percent + '%; 速度：' + speed + 'Mb/s;');
+            },
+            onFileFinish: function (err, data, options) {
+                console.log(options.Key + '上传' + (err ? '失败' : '完成'));
+            },
+        });
+    } else {
+        console.log('本次没有需要上传的文件');
+    }
+    
     // Note: 执行删除
-    cos.deleteMultipleObject({
-        Bucket: `${COS_BUCKET}`,
-        Region: `${COS_REGION}`,
-        Objects: removeFiles,
-    }, function(err, data) {
-        if (err) {
-            console.log('删除过程出现错误:', err);
-        } else {
-            console.log('删除结果:', data);
-        }
-    });
+    if (removedList.length) {
+        cos.deleteMultipleObject({
+            Bucket: `${COS_BUCKET}`,
+            Region: `${COS_REGION}`,
+            Objects: removeFiles,
+        }, function(err, data) {
+            if (err) {
+                console.log('删除过程出现错误:', err);
+            } else {
+                console.log('删除结果:', data);
+            }
+        });
+    } else {
+        console.log('本次没有需要删除的文件');
+    }
 }
